# docs/openapi/openapi.yml
openapi: '3.1.0'
info:
  title: Goofy Grove Server API
  version: '0.1.0'
servers:
  - url: https://127.0.0.2:3003/api/v1
security:
  - JwtToken: []
...
paths:
...
  /users/me:
    get:
      tags: [Users]
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        401:
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags: [Users]
      summary: Update current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          $ref: '#/components/responses/BadRequestError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        404:
          $ref: '#/components/responses/NotFoundError'
        default:
          $ref: '#/components/responses/InternalError'
  /users/me/avatar:
    post:
      tags: [Users]
      summary: Upload user avatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: Avatar image file
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        400:
          $ref: '#/components/responses/BadRequestError'
        401:
          $ref: '#/components/responses/UnauthorizedError'
        403:
          $ref: '#/components/responses/ForbiddenError'
        default:
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
      JwtToken:
        type: http
        scheme: bearer
        bearerFormat: JWT
        description: JWT token in Bearer format
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  data:
                    properties:
                      code:
                        const: unauthorized
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  data:
                    properties:
                      code:
                        const: internal_error
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  data:
                    properties:
                      code:
                        const: bad_request
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  data:
                    properties:
                      code:
                        const: not_found
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ErrorResponse'
              - properties:
                  data:
                    properties:
                      code:
                        const: forbidden
  schemas:
...
    UserUpdateRequest:
      type: object
      required: [avatar_uid]
      properties:
        avatar_uid:
          oneOf:
            - type: string
            - type: 'null'
          description: File uid to bind as avatar, or null to clear
    FileUploadResponse:
      type: object
      properties:
        error:
          type: boolean
          const: false
        data:
          type: object
          required: [uid]
          properties:
            uid:
              type: string
              description: Uploaded file uid (status Created until bound via PATCH)
...
    ErrorCodes:
      type: string
      enum:
        - unauthorized
        - forbidden
        - internal_error
        - bad_request
        - not_found

    # Responses
    ErrorResponse:
      type: object
      required: [error, data]
      properties:
        error:
          type: boolean
          const: true
        data:
          type: object
          required: [reasons, code]
          properties:
            reasons:
              type: array
              items:
                type: string
              example: ['Failed to authenticate user']
            code:
              $ref: '#/components/schemas/ErrorCodes'
...
    UserResponse:
      type: object
      properties:
        error:
          type: boolean
          const: false
        data:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      required: [uid, username]
      properties:
        uid:
          type: string
          description: User Unique ID
        username:
          type: string
          description: Nickname, login, username
        avatar_uid:
          oneOf:
            - type: string
            - type: 'null'
          description: Current avatar file uid